traductionLettre/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
appscript==1.2.5
asgiref==3.7.2
certifi==2024.2.2
cfgv==3.4.0
click==8.1.7
contourpy==1.1.1
coverage==7.4.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
document==1.0
docx==0.2.4
docx2pdf==0.1.8
et-xmlfile==1.1.0
Faker==22.6.0
fastapi==0.109.0
filelock==3.13.1
fonttools==4.43.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
identify==2.5.33
idna==3.6
kiwisolver==1.4.5
loguru==0.7.2
lxml==5.1.0
matplotlib==3.8.0
networkx==3.1
nodeenv==1.8.0
numpy==1.24.2
openai==1.12.0
opencv-python-headless==4.7.0.68
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pdfrw==0.4
pillow==10.2.0
platformdirs==4.2.0
plotly==5.17.0
polyfactory==2.14.1
pre-commit==2.15.0
promise==2.3
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
pydot==1.4.2
pygame==2.1.3
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.35.1
tenacity==8.2.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.3
uuid==1.30
uvicorn==0.27.0.post1
virtualenv==20.25.0
whitenoise==6.6.0
